requirements.txt
# AI-Powered HTTP Request Analyzer – Dependencies

# LangGraph / LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0

# Web API
flask>=3.0.0
flask-restx>=1.3.0

# Core
pydantic>=2.0
python-dotenv>=1.0.0
typing_extensions>=4.0

# Middleware clients
redis[hiredis]>=5.0.0
qdrant-client>=1.9.0
transformers<4.45.0
sentence-transformers>=3.0.0
fastembed>=0.3.6
aiohttp>=3.10.0
datasets>=2.20.0
gatekeeper_ml
